pyproject.toml
[project]
name = "google_functionregistry"
version = "0.2.6"
description = "Google Function and Parser Registry"
authors = ["Torstein Sørnes <t.soernes@gmail.com>"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "betterpathlib>=0.1.1",
  "python-dotenv>=1.0.1",
  "pydantic>=2.9.2",
  "google-genai>=1.4.0",
  "six>=1.17.0",
  "ipython>=9.0.1",
  "google-generativeai>=0.8.4",
]
